+++ scripts/helperFunctions.js
@@ -419,7 +419,7 @@ export let chris = {
         return DAE.convertDuration(item.system.duration, chris.inCombat());
     },
     'getCriticalFormula': function _getCriticalFormula(formula) {
-        return new CONFIG.Dice.DamageRoll(formula,    {}, {'critical': true, 'powerfulCritical': game.settings.get('dnd5e', 'criticalDamageMaxDice'), 'multiplyNumeric': game.settings.get('dnd5e', 'criticalDamageModifiers')}).formula;
+        return new CONFIG.Dice.DamageRoll(formula, {}, {'critical': true, 'powerfulCritical': game.settings.get('dnd5e', 'criticalDamageMaxDice'), 'multiplyNumeric': game.settings.get('dnd5e', 'criticalDamageModifiers')}).formula;
     },
     'getSize': function _getSize(actor, sizeToString) {
         let sizeValue;